requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pyyaml>=6.0

# NLP & ML libraries
nltk>=3.6.0
transformers>=4.18.0
torch>=1.10.0
bert-score>=0.3.11
rouge-score>=0.1.2
spacy>=3.2.0

# Data handling
Pillow>=8.3.0  # For image processing in future stages
opencv-python>=4.5.0  # For image & video processing in future stages

# Web app
streamlit>=1.15.0

# Evaluation & utilities
tqdm>=4.62.0
ipykernel>=6.0.0  # For notebooks